requirements.txt
streamlit>=1.32.0
tweepy>=4.14.0
pandas>=2.0.0
plotly>=5.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
